Add ability to define callbacks in ActionMailer using AbstractController::Callbacks.
Prior to this commit, there isn't a good way of adding things like
default inline attachments to an email.  This Stack Overflow thread
shows people using hooks like the 'default' method in ActionMailer::Base
to call a Proc for message configuration:

http://stackoverflow.com/questions/5113121/rails-use-same-attachment-for-all-emails-using-layout

This has the unintended side effect of setting a message header, so it's not a good solution.

This pull request adds support for message modifications by including AbstractController:Callbacks
in ActionMailer::Base. It includes tests and documentation for the functionality
provided by including this module.