Prevent not-intended loading of `ActionDispatch::IntegrationTest`
After ac9e74c3495dbd9aa7c9147878a4abbe7e0169d5 `ActionDispatch::IntegrationTest`
class is loaded and defined in all Rails environments, not only test but also
production. This is not-intended loading of a class which is only used in
test environment.
To prevent not-intended loading, add `ActiveSupport.run_load_hooks` to
`ActionDispatch::IntegrationTest` with `action_dispatch_integration_test` name
and use it in `ActionMailer`.